Refactor authentication imports and enhance sign-in flow; remove unused signup form component
