+++ prisma/schema.prisma
@@ -53,6 +53,7 @@ model User {
     image         String?
     accounts      Account[]
     notes         Note[]
+    sessions      Session[]
 }
 
 model VerificationToken {
@@ -61,4 +62,12 @@ model VerificationToken {
     expires    DateTime
 
     @@unique([identifier, token])
+}
+
+model Session {
+    id           String   @id @default(cuid())
+    sessionToken String   @unique
+    userId       String
+    expires      DateTime
+    user         User     @relation(fields: [userId], references: [id], onDelete: Cascade)
 }
