+++ src/app/page.tsx
@@ -1,7 +1,8 @@
 import Link from "next/link";
 
 import { LatestPost } from "~/app/_components/post";
-import { auth } from "~/auth";
+//figure out which import to use /auth /server/auth
+import { auth } from "~/server/auth";
 import { api, HydrateClient } from "~/trpc/server";
 
 export default async function Home() {
